requirements.txt
# 14. requirements.txt
Flask==2.3.2
flask_restx==1.1.0
selenium==4.15.2
pandas==2.1.1
PyJWT==2.8.0